Enum Level extrait de QuestionBank car utilisé en dehors de la classe.
